Schedule performance improvements from BC's site
These were local changes on BC's site; I don't know who made them, but these
are the parts that it looks like we want and don't already have.  The timeslots
thing should help a little; the tag matters less since it's cached but we might
as well do it less stupidly.