Remove `ESPUser.__olduser`
When converting a `django.contrib.auth.User` to an `ESPUser` using
`ESPUser(user)`, we were previously keeping around a reference to the
underlying `User` object as an attribute on the new ESPUser.  While this might
be nice in theory, when the user was wrapped in a
`django.utils.functional.SimpleLazyObject` (as, for instance, is
`request.user`), pickling it would raise an error due to a django bug
(https://code.djangoproject.com/ticket/25389).

Luckily, we appear to only be using the reference to the underlying user in a
few cases, so we can just work around the bug in those cases.  One is
`ESPUser.is_authenticated`, which manually calls the underlying `User` method.
This is not necessary; we have already copied over the `User`'s `__dict__`.
The other is `__eq__`, which is overridden to make an `ESPUser` equal to its
corresponding `User`.  This is no longer necessary; django does this for all
proxy models as of 1.7.  Finally, on calling `ESPUser()` on an `ESPUser`, the
underlying user was being used to populate the new `ESPUser`.  This is also
unnecessary; we can just copy from the `ESPUser` instead.

This change is a little scary, because of all the `__dict__` manipulation going
on.  But I think it if anything simplifies things, and also works around the
aforementioned django bug.  (Note that pickling and unpickling an
`SimpleLazyObject` wrapping any model may still raise a warning due to the same
bug, but at least it will work.  Right now it's just skipping a safety check,
so it's not a big concern.)