Purge Django's app cache after custom forms tests.
Since each test runs in its own database transaction, any tables
created during the custom forms tests are not visible to subsequent
tests. However, there are various places in Django's app loader where
lists of models and fields are cached, so it's possible for another
test to do something (such as User.delete()) that gets the models from
the cache and try to run queries against them, causing an error when
the corresponding tables no longer exist.

Add a new method to purge the app cache for a custom form, and call it
when deleting a table and when tearing down the custom forms tests.