ajax scheduler: don't allow scheduling on non-contiguous timeslots.
AJAX scheduler did not have check for non-contigous timeslots
(e.g. 5:05 pm to 5:55 pm, and then 7:05 pm to 7:55 pm). We now
detect non-contiguity by assuming the minimum time difference
between classes is the expected "passing period", and failing
if the time difference on the selected timeslots exceeds this
expectation. (Hardcoding a value like 20 minutes is not
guaranteed to work, e.g. timeslots may not be one hour long.)

Fixes #1797