AJAX scheduler: always show top siderbar section.
The top sidebar section includes the unschedule button and selected class details.
The current behavior is that the entire right sidebar is a scrollable div, which
means that when you scroll down to select a class, the top sidebar section is no
longer on the screen. So if you later select a class to unschedule, you need to
scroll all the way back up. That is no good.