AJAX scheduler: allow locking classes during scheduling.
Adds a set scheduling comment dialog that allows admins to
set a message when scheduling classes. A lock option can also
be checked to lock the class into the assignment. If a section
is locked, then it cannot be unscheduled or rescheduled (this
restriction is implemented client-side).

Implements part of Issue #460.