Remove StatsMiddleware
This appears to have been a way to time and print SQL queries, long before we
had the debug toolbar.  It still works (!), but the formatting is pretty
broken, and the debug toolbar is way better, so we can remove it, along with
the DISPLAYSQL setting that was used to turn it on or off.