Rewrite program cap logic
This change rewrites the program cap logic, to support new features requested
by Stanford.  In particular, rather than using `Program.isFull()` or
`Program.program_size_max`, all callers should now use
`Program.user_can_join(user)`, which will correctly determine if a user is
already in the program (in which case they can always register for it), and if
not, whether the cap is full.  Module views can use `@meets_cap` which will
display an error page if the user can't join the program.  (The cap can also be
overridden by giving them the "OverrideFull" permission.  Previously this was
implied by "Student/All"; it is no longer.)

In addition, programs wishing to set a per-grade program cap (this was
Stanford's request) may set the program_size_by_grade tag, which is documented
in `Program.user_can_join` and overrides `program.program_size_max`.

This also makes the cap apply to most program modules.  Previously users would
only encounter it when they got to the point of trying to add a class; this is
a bad user experience at best, and meant it was impossible to use a cap for the
lottery.  Now the program cap is checked before you do anything.  If we later
find it necessary to make this more configurable, we can do so at that point.

One thing where this doesn't update the logic is in the lottery script itself.
This should be fine for now, since the cap will prevent more students from
signing up than should be allowed.  If necessary we can add the cap to the
lottery, but that would be difficult.

This also fixes a minor bug: previously the lunch constraints module would not
act correctly if no program cap was set.  (It would set the lunch class's size
to zero.)  Now it will do something reasonable.