Fix fab update_deps
The manage.py command doesn't work when you're installing for the
first time and need to create the virtualenv, or your Python
dependencies are otherwise too out of date for Django to start. Run
the update_deps.sh script directly instead.

The interaction between fab and apt in the terminal output is a bit
ugly, but oh well.