Remove allow_guardian_no_email and require_guardian_email tags
These were only in use on the Chicago site, which also disables the guardian
profile form entirely (via local changes and template overrides), making them
moot.  Their behavior was also confusing: require_guardian_email would set the
guardian email field to be required, while allow_guardian_no_email would add
another checkbox which, if checked, would ignore that requirement.  This is
silly, since the field is not required and there doesn't seem to be demand to
make it so.  If there is, we can add a simpler setting at that point.