Remove ask_student_about_post_hs_plans tag
This tag and the associated data appear to have been used by Chicago several
years ago, but are not used on any sites now.  Since they have data in it, I
left the field there at least for now, but we can see if they actually want
that data, and if not, remove it.