Refactor some code around creation of ESPUsers
The immediate reason for this was removing the code which can cast
a Django User or AnonymousUser to an ESPUser, which caused a nasty
mess recently. Making a new auth backend which returns an ESPUser
instead of a User was simple enough. However, AnonymousUser means
there are still some issues.

Previously, ESPUser inherited from both User and AnonymousUser. We're
not sure if inheriting from AnonymousUser was doing anything, but it
was there. Either one would get cast to ESPUser, which would add
a property indicating whether it was anonymous. This means that we're
used to having all of ESPUser's methods available even for anonymous
users, and we don't know to what extent we're using that. It's a
pretty weird way of doing things, though.

So instead, I've put all the methods in a base class, and created
our own version of AnonymousUser. ESPUser and AnonymousESPUser now
inherit from both BaseESPUser and the corresponding Django class.
Any methods which override User will have to go in ESPUser directly,
but so far we've only been doing that to handle anonymity, which is
no longer needed.