Remove esp.utils.defaultclass
This was somehow intended to allow for alternate implementations of classes.
Only one class was set up to use it (StudentProfileForm), and as far as I can
tell we just always used the default implementations, even on chapter-specific
branches -- Chicago just edited their version of StudentProfileForm rather than
using the defaultclass thing.