Prevent people from generating many student schedules as PNG
Generating PNG schedules for a lot of students will cause
`convert` to use a huge amount of memory and make the server sad.
It also doesn't work, since we just return the first page of the
PNG anyway.  So don't let people do that.  This has caused issues on both
diogenes and esp.mit.edu recently.  Fixes #1874.