Fix ajax autocomplete css issue
Grappelli apparently changed their css in a way that depends on having any
form widget's last tag be an `input`, `select`, or `textarea`.  (In particular
they have some css that's roughly `input + .class-name, ... { ... }`, where
`.class-name` is for the plus button that goes next to `ForeignKey` fields.)
We previously had a `script` tag last.  In order to move it first, I had to
also wrap it with a `$j(function () {});` so it won't run before the elements
to which it refers exist.