Ignore constraints when removing classes as onsite morphed user.
The backend already ignores constraints when adding classes as onsite
(the frontend displays the would-be error on the fillslot page, and
gives the volunteer a strictly-worded "are you sure?" prompt before
overriding it). I think such a warning is less necessary for removing
classes, and this fixes a situation where it's possible for an onsite
user to override constraints on multiple classes and then not be able
to undo their actions.