Remove unused template blocks
It seems that a long time ago, we had a block in `main.html` or the
then-equivalent called `subsection` and/or `subsection_name` which templates
were expected to extend.  This is long gone, but due to copypasta the block
still remains in 100+ templates.  This commit removes it, along with loads of
the template library `subsection`, which was only used therein.