Remove unused markup template filters, and move to utils
We only use markdown, we know it's installed and what version we have, and we
don't use its extensions, so we can remove a bunch of this code and greatly
simplify the remainder.  If we want to do something more complicated, there are
probably packages out there that will do it for us.

I also moved them to utils, because that's what they are.