Clean up uses of array_agg()
Apparently postgres' array_agg() returns null instead of empty array
if there are no rows, and its various consumers in our code deal with
this in different ways (one in Python, one in JS as of 07811b6, and
one just assumes it never happens).

The ARRAY() constructor does not have this problem, so use that
instead. Also remove an array_to_string() call that is useless because
it just gets split into an array again anyway.