Allow editing queries in query builder (and class search)
Previously, on submitting the query builder, you would be presented with an
English representation of your query, along with the results.  Now you'll be
presented with the filled-in query builder, along with the results.  This means
we don't have to wade into the mess that is trying to construct vaguely English
representations of queries, and makes it easy to modify your query and try
again.  Fixes #1620.