Rewrite ensure_required_transfers
Since transfers aren't being executed, this function deletes and recreates a
user's required transfers (program admission) every time they load certain pages
(paynow_cybersource). This causes the transfer's ID to change, which makes it
hard to keep track of. The reason we do this is so that if the program's cost is
changed, users haven't paid but who we've calculated an amount due for get an
updated amount due.

Right now, once ensure_required_transfers is called, it locks in the price of
admission. Later, once we have a way of actually tracking which transfers have
been executed, this will be fixed.