Add a lint test
Now that we don't have any undefined names (see #1944), let's keep it that way.
This adds a dependency on pyflakes and a test to check that it doesn't find any
undefined names.  We can add more similar tests as we see fit; it should be
fairly trivial to add more pyflakes-based ones and not that hard to use other
linting libraries.