Cybersource: properly autofill first name and last name fields
Right now, the user's first and last names are joined and put in the "first
name" field. In reality, we should put the first name in the "first name" field
and the last name in the "last name" field. We should also make sure that both
the first and last name fields are required.