Cybersource: use HTML5 input types
This makes the browser validate that the email address looks like an email
address and also makes the telephone number field open a numeric input pad on
mobile (in case anyone is filling this out on mobile). The email address
validation is particularly important, as Cybersource will instantly fail when
passed a malformed email address.