Update new_site.sh some more
The changes in #1529 did the lion's share of the work, but there were some more
updates that were needed to bring the script in line with current practice.  I
also added a few steps for things that were done manually but don't need to be.

At this point the only things the script doesn't do are making the commit to
the /etc repo (because the human may wish to check things first, or to separate
out any preexisting changes from theirs) and updating the DNS (since that
happens on a different server; the script gives instructions).