Add deployment docs
Since `new_site.sh` does most things now, these are fairly short.  I also added
a TODO for having docs for how to set up a new server; it may not be worth
trying to be too exhaustive about that, but it would be nice to at least have a
bit of an idea, because we're going to have to do it again some day.