Convert print statements to logging
Also `warnings.warn`.  I didn't convert useful_scripts, since in a lot of those
it's not worth trying to do real logging, and in some it would muck up the
output.  I also converted a number of `if debug: print x` to just
`logger.debug(x)`.

A lot of this was generated using several `codemod.py` commands; some was done
manually.