Remove ProgramModuleObj's special __getattr__
This is the key change to clean up the module extension system: you can no
longer access attributes of module extensions on the `ProgramModuleObj`.

This has been deprecated for a while, but unfortunately in our current config,
deprecation warnings don't go anywhere useful.  But I grepped through the
codebase for each attribute of each module extension, and (other than the
`get_program` mess removed in 1d3587f) was unable to find any users that were
not obviously getting the attributes from the module extension, rather than the
module.

This is still a bit scary, because I could have missed something, or there
could be something ungreppable.  But I tried to be careful, and I think this is
worth doing.  If we later find any issues, they should be easy to fix at that
point.

This will allow a bunch more cleanup, which will come in a separate commit.