Add program field to module extensions
This will replace the module field.  The population is a little sketchy on old
programs where there are sometimes multiple module extensions for the same
program (including sometimes ones with the wrong module set).  But that was
undefined behavior to begin with (it depended on the order the database
returned them in when no ordering was specified, and whether you used the ones
with the wrong module depended on who was asking), and only existed on old
programs, so I'm not too worried about it.