Update all callers to refer to the program rather than the module
The `module` field of a module extension wasn't that meaningful anyway, apart
from as a way to get to the program, so updating all of its callers was fairly
easy.  It's still a little scary, because there are a lot of ways to refer to a
module field, but I think it will be okay, and I'm going to leave behind a
property for places where that works.