Clean up module extension creation
Instead of creating the module extension when instantiating the module via
`getFromProgModule`, we'll create it with a signal attached to the
`Program.program_modules` M2M.  This ensures that the module extension will
exist any time the module does.

We could just autocreate them on program creation to fully decouple them from
program modules, but that seems like overkill; the dependency is loose enough
that it should be easy to change when we instantiate them later.