Remove some unused tags and TeacherInfo fields
These fields were only used if you the tag "teacherinfo_reimbursement_options"
set.  The only chapter that had it set was Chicago, where this appeared to be
in error -- it was set to " " (which in this case is truthy) -- and in any case
the fields were manually excluded using teacher_profile_hide_fields.  It looks
like they last used it in 2011.

It seems like reimbursement systems are idiosyncratic enough that chapters will
need to either use a teacher custom form for ask for whatever information they
need, or just ask for it out-of-band or in an existing field.  (I think most
chapters ask for it out-of-band.)  So it's not worth us having shared
functionality for it at present.