Deprecate and redirect old teacher bio URLs
The canonical URLs for bios are /teach/teachers/<username>/bio.html.  But it
turns out you can also access them at
/teach/teachers/<lastname>/<firstname><num>/bio.html and at both of those URLs
with /learn/ instead of /teach/.  This commit makes the deprecated URLs
permanently redirect to the correct ones; then sometime in the future we can
just remove the old ones and simplify the code a good bit.