Fix indentation bug introduced in 4fbf15658af69085
If you look at the diff, the body of ajax_schedule_assignreg was copied
over from the old ajax_schedule_class, but an extra level of indentation
was accidentally introduced after the for-loop after classrooms.

This was mostly okay because the body of the for-loop would only ever
run once (because it returns at the end), with the first classroom in
classrooms, which is what the intended code does anyway as long as all
the classrooms are the same, and all that the original for-loop did was
check that all the classrooms are the same anyway.