Fix shebang lines
Several of our `#!` lines are `#!/usr/bin/python` which doesn't do the right
thing if you are in a virtualenv.  (In particular, most of the scripts will
auto-activate the virtualenv if you run them from outside of any virtualenv,
but if you run them from inside the virtualenv but with /usr/bin/python as your
interpreter, they may do the wrong thing.)  This changes them to use
`#!/usr/bin/env python` which is more universal.