Refactor of user_classes.js
In order to decrease the amount of custom code needed for the way different
themes handle the login box, I did a refactor of `user_classes.js`.  Now,
instead of prefilling the entire login box with a blob of HTML for logged-in
users, we'll inject individual user variables into specific `<span>`s.

For sites with the themes, this should not be a functionality change.  For
sites with custom themes, we'll need to adapt their template overrides.  This
will consist of replacing the `div#loginbox_user_name`, if they use it, with
`{% include "users/loginbox_userinfo.html" %}`.  For sites that don't use it,
we'll want to migrate their templates to the new system to remove some of their
custom javascript, but it will work as-is.  This will also simplify things when
we add the `berkeley-alpha` theme.

I also moved some `barebones`-specific bits of JS from the shared
`user_classes.js` to a theme-specific script.  Finally, I moved the binding of
`update_user_classes` to document ready into the script itself rather than
having it in a JS snippet in `elements/html`, along with some other stylistic
cleanup.  None of this should involve functionality changes.