Cybersource: simplify the payment page
Entering invalid information (e.g. an incorrect state abbreviation) causes
Cybersource to return with an ERROR code. Because we auto-fill the country as
'US', this makes it impossible for non-U.S. students to pay. The simplest way to
solve this problem is to get rid of all the fields and have people enter this
information directly into Cybersource.