Clean up our URL patterns (#2121)
The main goal of this commit is to remove uses of patterns() and string
view arguments to url(), both of which are deprecated. While I was doing
that, I moved some patterns into per-app files and cleaned up the main
urls.py a bit. Progress towards, but not a complete fix for, #1292.

(I still compulsively fix warnings.)