Refactor and simplify profile handling
Some miscellaneous refactoring and simplification, in preparation for splitting
the profile module.

This might fix #1832 (thanks @mgersh for noticing); it seems like we intended
to clear the email address to make you make sure to update it, but that wasn't
actually working and I don't think it's desired behavior these days anyway.