Allow hiding the FAQ link in contact info
This was requested by USC, and seems pretty reasonable -- we already allow
hiding the FB link.  In this case, we want the default to exist, including for
sites that haven't updated their theme settings since this was added.  So we'll
fill it in as `/faq.html`, if it's not set, and thereafter display it only if
it's nonempty.

Really, we should probably have a more complete editor here, like we do for the
site tabs, but I've just added a TODO for now.  (So this is only a partial fix
for #2156.)