Update Makefile to write the DB dump with permissions
The current user doesn't always have permissions on the site directory,
so they couldn't write the DB dump.  Now we'll do this with sudo.  I had
to get a little fancy, and make sure we ask for permissions *before* we
start doing fancy things with pipes, but it should work; tested on USC.