Try using create() instead of get_or_create() in dbmail cron
Hopefully this will solve our problems with Stanford's emails.  We don't want
to do it in general, since we'd lose deduping, but this email it doesn't matter
since it's already a mail merge.  So if this works we'll figure out how to do
it more generally.

(cherry picked from commit 38c2b41d19056e6cc9ef93557aa1846b2ab5d879)