Fix queryset batching for dbmail_cron
This is a fix to 64f93757.  I realized since we're marking the emails as sent
every time, and this method of batching actually reruns the query, we should be
grabbing the *first* batch every time, rather than moving our slice forward.
(It doesn't super matter if we're perfect, since the next dbmail_cron will
retry, but we may as well do it right.)