Consolidate qsd hidden classes, add .qsd_default_hidden
Use one fewer CSS declaration. More importantly, avoid a sketchy
operation in user_classes.js where

    $j(".qsd_bits").removeClass("hidden");

is not supposed to unhide this particular `<div>` because it only
removes one copy of the 'hidden' class and there's another one.
Unsurprisingly, this stops working in later versions of jQuery, which
would be nice to have eventually.