Fix .less regexes: don't include variables_foo.less twice (#2245)
The code tries to include all variables*.less first and then all other
files second, but the regex used to match the latter type of file name
isn't correct and inverting a regex is too damn hard to do or verify
anyway. Grab all the *.less files in one pass and manually sort out
which ones start with 'variables' outside.