Add profile affiliation field (#2155)
This change is requested by Stanford, to add an affiliation field to the profile form. The options are undergrad, grad, postdoc, other, none ("none" is not needed by Stanford but is added so that this change can be used at all schools). There is an additional field to specify further, if "other" or "none" is selected.

I also removed several questions from the form that are made redundant by this change: from_here, is_graduate_student, school. The fields are not removed from the database, because they are used elsewhere. Instead, these fields are populated automatically based on the affiliation.

"Affiliation" was also added as a display column and filter on the teacher infos admin page, and also on the userview display.