AJAX scheduler: make teachers with no availability always available
If a teacher is not available for any timeslots, consider them available
for all timeslots, with a different striped yellow-green style to warn
about the fact that this was used. Fixes #1005 probably?

(If the availability module is enabled, scheduling classes onto such
fallback timeslots will fail on the server side and be reported to the
client. But I think this is fine since it moves us closer to #1401?)