New Phase Zero (pre-lottery) module (#2190)
This implements a new lottery system to determine which students can enter the
program, along with an override system and a script to run the lottery.
Students may enter the program in groups with other students (e.g.  siblings).
It integrates with existing grade-based program cap functionality to accept a
certain number of students from each grade.

Requested by Stanford Splash.  Fixes #2350.