Fix theme tests
Adding the bigpicture theme caused a conflict between the tests for it and for
fruitsalad -- the two both set the same default files, which causes us to hit a
confirm page to clobber them in the test, and we don't know to click that, so
things don't work right.  This could also happen even before bigpicture, if
running tests locally on a dev server that had such a conflicting file.

One way to fix it would be to click past the confirm page. But that's a problem
since it may be clobbering real files, if someone is running locally!  So
instead, we now back up the offending files, and restore them at the end.  This
is super hacky -- we really shouldn't be modifying the user's files at all ever
if we can avoid it -- but it at least fixes the problem for now.

Fixes #2388.