ClassChangeController: work around overflowing classes
Take note of the "badness" of the assignment, which is how much it
overflows true class capacities, and add a "pessimism" parameter that
makes the algorithm more conservative about guessing that everybody who
wants to switch out will get to. This way the lottery can be tuned to
not overflow classes. This is a hacky workaround to patch the confusing
student-kicking algorithm applied to "fix" the greedy student-assigning
process, which doesn't always manage to get enrollments below true class
capacities.