Add duration-from-now labels next to deadline datepickers
On datepickers on the deadline page, add a little colorful Bootstrap
label with a glyphicon that indicates how far away the selected time is
from now, which will alert users about dates more than 120 days in the
past or future and otherwise indicate whether the date is in the past or
future.

Also make datepicker textboxes narrower so the labels are less likely to
wrap.

Fixes #1700.