Add per-section observers to model, zeroth draft
This displays the observers in one's classes and some limited data about
the sections one is observing, although it contains no front-end for
adding observers or any schedule checks whatsoever.

A notable change orthogonal to observers is that section data is shown
for Unreviewed classes; this may not be desirable in general and we
should discuss it more, but is needed for Spark.

Related to #1990; I couldn't justify naming the thing I added
"assistant" or "TA", though, since those concepts seem more likely to be
per-class, while observers per section make sense. Generally "observer"
assumes less about the role, I think.