Display better info (#2523)
Added a grade filter to the scheduler. This takes the minimum or maximum grade inputted and only displays sections that have that grade as the lower or upper bound. For example, if you set the min to 10, you will get 10-11 and 10-12th grade classes, and if you also set the max to 12 you will get just 10-12th grade classes.

Added a column to show whether the user is activated or deactivated in the user search. This is helpful if one person has many accounts.

Added a select all button to the recipient checklist in the communications panel.