Fixes to Makefile (#2518)
The main fix here, to #2508, is to `set -o pipefail` where we have a
nontrivial pipeline, so that if the first command (the `pg_dump`) fails,
at least we'll obviously crash.  In the past it would just silently fail
and continue working, which is bad since then we have no db dump.

While I was in the area I also converted some tabs to spaces, because
`make` is really just happier with tabs.