Fix ordering of two-phase priorities (#2558)
Fixes #2552.

I wasn't able to replicate the odd behavior reported in #2552 on my dev server, but I replaced the old sorting behavior with a much better way of parsing through the priorities (using dictionary keys), which seems to be better nonetheless. It also fixed another bug that was introduced with #2425 where the priorities would shift upwards (in priority number) if there were skipped priority levels (in other words, if a student had set Priority/1 and Priority/3, they would be displayed as Priority/1 and Priority/2). This is also fixed by the new implementation.