Add dates to volunteer offers (#2586)
Volunteer requests are now separated by date on /volunteering. I changed the displayed time in the first column so that it wasn't redundant with the new headers. While I was at it, I fixed up some of the admin page displays so that they now include the dates of the events.

Fixes #2465.