Add some logging to dbmail (#2627)
If the script is working, it logs a bunch.  But if we crash, we were not
logging where we were before the crash.  Also, due to a logging
misconfiguration, the logs weren't going anywhere (because the logger
name didn't start with `esp.`).  Now it should all work!

Fixes #2624; slightly adapted from bd3c173 which is on yale-sr10.