Fix for deleted users when users are deleted
Custom forms were breaking if a user associated with a response was deleted. We probably shouldn't be deleting users anyways, but just in case this happens again, this is a fix for that (we just skip including user data in the response in such cases, but we can still fill in all of the other information).

Fixes #2673.