Teacher big board fixes (#2685)
The teacher big board could break during the calculation of the hours if there was a class with no sections. I've fixed that and a bug that I think I reported on the old slack channel but never here. The class-hour lines would stop abruptly on the plot because they excluded lunch classes but the other calculations didn't. Now everything excludes lunch classes.

Fixes #2662.